+++ src/services/article/hooks/useArticleCategories.ts
@@ -1,0 +1,12 @@
+import { useQuery } from '@tanstack/react-query';
+
+import { queryKeys } from '~/react-query/common';
+import { getArticleCategories } from '~/services/article/apis';
+
+export const useArticleCategories = () => {
+  return useQuery({
+    queryKey: queryKeys.articles.categories(),
+    queryFn: getArticleCategories,
+    select: (categories) => categories.sort((a, b) => a.boardId - b.boardId),
+  });
+};
