+++ src/services/article/hooks/useCreateArticle.ts
@@ -1,0 +1,9 @@
+import { useMutation } from '@tanstack/react-query';
+
+import { createArticle } from '~/services/article/apis';
+
+export const useCreateArticle = () => {
+  return useMutation({
+    mutationFn: createArticle,
+  });
+};
