+++ src/services/article/hooks/useRemoveArticle.ts
@@ -1,0 +1,9 @@
+import { useMutation } from '@tanstack/react-query';
+
+import { removeArticle } from '~/services/article/apis';
+
+export const useRemoveArticle = () => {
+  return useMutation({
+    mutationFn: removeArticle,
+  });
+};
